+++ CHANGELOG.md
@@ -13,10 +13,10 @@ Changelog
 
 ### Added
 
-- Added `Assert::notNegativeInteger()` and `Assert::negativeInteger()`
 - All assertion methods now return the checked value.
 - Added `notInArray` and `notOneOf`.
 - Added `isInitialized`, to check if a class property is initialized.
+- Added `notNegativeInteger` and `negativeInteger`
 
 ### Fixed